move trace_object to mallocspace
